+++ example/test/widget_test.dart
@@ -1,16 +1,10 @@
-// This is a basic Flutter widget test.
-//
-// To perform an interaction with a widget in your test, use the WidgetTester
-// utility that Flutter provides. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
-// tree, read text, and verify that the values of widget properties are correct.
-
 import 'package:flutter_test/flutter_test.dart';
 import 'package:qrscan_example/main.dart';
 
 void main() {
-  testWidgets('Verify Platform version', (WidgetTester tester) async {
-    // Build our app and trigger a frame.
-    await tester.pumpWidget(const MyApp());
+  testWidgets('one Scan button', (WidgetTester tester) async {
+    await tester.pumpWidget(const DemoApp());
+    expect(find.text('扫描'), findsOneWidget);
+    expect(find.text('点扫描'), findsOneWidget);
   });
 }